Pagos con PayU y MercadoPagos

+++ laravel/database/migrations/2018_03_06_113859_create_sliders_table.php
@@ -17,7 +17,7 @@ class CreateSlidersTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('slug');
-            $table->smallInteger('priority')->unsigned();
+            $table->tinyInteger('priority')->unsigned();
             $table->string('url');
             $table->string('main_text');
             $table->string('small_text')->nullable();
