+++ laravel/database/migrations/2018_03_08_174116_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->increments('id');
             $table->integer('buyer_id')->unsigned();
             $table->foreign('buyer_id')->references('id')->on('users');
-            $table->smallInteger('status')->unsigned();
+            $table->tinyInteger('status')->unsigned();
             $table->timestamps();
         });
     }
